Adds the XM51 Breaching Scattergun (#5086)
[Design doc (a bit out of
date).](https://hackmd.io/uVhllc-KQSGcvjhEUw66bw?edit)

The XM51 breaching scattergun is a 16-gauge, pump-action mag-fed shotgun
that utilizes specialized breaching shells. It allows marines to break
through doors and walls alike with ease at a safe distance. Each shell
fires out four groups of pellets, the main one dealing 55 damage and the
other three dealing 10 each. Though it has decent damage, this weapon
performs weakly in direct combat compared to other options and is
entirely designed for a support role. Its maximum effective range is
about 4 tiles.

It takes:
- 1 shot for doors
- 2 shots for regular resin walls
- 3 shots for thick resin walls
- 2 shots for regular walls,
- 4 shots for reinforced walls.
- 4 shots for pylons,

Each XM51 kit comes with a dedicated holster for the weapon, along with
two spare magazines, two boxes of shells and a stock. Being able to wear
the weapon around your waist gives the user a great deal of flexibility
when it comes to their loadout. The stock makes it unable to be
holstered, but it does give it a very devastating two shot burst-mode
that quickly clears out walls.

Ammo for the weapon is a bit scarce but still present in a decent
quantity. Requisitions starts with a single 120 round shell box which
they can purchase more of, along with a couple of spare magazines. More
magazines can be purchased, and ASRS has a chance of sending up spare
ammo as well.

This weapon is based on the [actual M51 breaching
scattergun.](https://avp.fandom.com/wiki/M51_Breaching_Scattergun)

Marines are very cowardly when it comes to pushing. Especially so when
they're facing countless thick resin walls and doors. This contributes
to a stale meta where the best option is to just sit behind cades and
wait for the other team to make a mistake, because why would you risk
dying to take down one resin door? That's not fun for anyone. Marines
have a few ways of making some breathing room for themselves from a safe
distance, with one of the best ones being the U7 underbarrel shotgun.
It's a very good and fun attachment, even putting aside the combat
capabilities it has.

However, it's entirely exclusive to the M41A rifles. Any marine that
wants to experiment with their weapon choice a bit and forgo the rifle
lose one of the best tools they have access to. The XM51 gives marines
another, more effective way to breach resin bunkers. In exchange for the
ease of use, abundance of ammo and good combat capabilities of the U7,
you gain a safer, faster, flexible and more potent tool.

<details>
<summary>Screenshots & Videos</summary>

https://github.com/cmss13-devs/cmss13/assets/17518895/6f4441d2-740e-4110-a9a6-32c39c424576

combat capabilities vs the u7

https://github.com/cmss13-devs/cmss13/assets/17518895/c0cb1359-cfd0-4b86-8c77-3885c86c5c8c

https://github.com/cmss13-devs/cmss13/assets/17518895/834d9bbb-3aee-4203-93a7-af40e28b7805

![xm51camo](https://github.com/cmss13-devs/cmss13/assets/17518895/1ada72ac-2b4a-441b-ac26-c0152a0eb62b)

![xm51ammo](https://github.com/cmss13-devs/cmss13/assets/17518895/471c2654-c242-4440-8b67-8e19bad8eb45)

![xm51statnostock](https://github.com/cmss13-devs/cmss13/assets/17518895/a672ec8a-3401-4830-a1db-821198dabf91)

</details>

:cl:
add: Added the XM51 breaching scattergun, available in small quantities
as a restricted weapon within Requisitions.

/:cl: